requirements.txt
# Client dependencies
pygame==2.5.2
websocket-client==1.6.4
Pillow==10.1.0

# Server dependencies
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1